pyproject.toml
[tool.poetry]
name = "langchain-practice"
version = "0.1.0"
description = ""
authors = ["sh1un <a0923183408@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.14"
langchain-anthropic = "^0.1.4"
python-dotenv = "^1.0.1"
boto3 = "^1.34.75"
langchain-openai = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
